Removed unused images and refactored stucture
