+++ README.md
@@ -1,20 +1,20 @@
 # OSRS automation
 
-This is education project for learning [OpenCV-pyton](https://github.com/opencv/opencv-python) library.
+This is education project for learning [OpenCV-python](https://github.com/opencv/opencv-python) library.
 
 I develop some automations for [Old School RuneScape](https://oldschool.runescape.com/) game.
 
-![Hello World](/documentation/hello_world.png)
+![Hello World](/readme/hello_world.png)
 
 ## Game settings
 
 You need to setup your game:
 
 1. Setup game client layout to **Resizable - Classic layaout** in interfaces settings;
-![Game Client Layout](/documentation/game_client_layout.png)
+![Game Client Layout](/readme/game_client_layout.png)
 
 2. Setup interface scaling to **150%** in interfaces settings;
-![Interface Scaling](/documentation/interface_scaling.png)
+![Interface Scaling](/readme/interface_scaling.png)
 
 ### High Level Alchemy
