+++ osrs/compare.py
