+++ readme/game_client_layout.png
