+++ tests/test_small_screenshot.py
